feat: add `decode_debug` to `ParamType`
