+++ packages/fuels-core/src/codec/abi_encoder.rs
@@ -59,6 +59,7 @@ mod tests {
     use crate::{
         codec::first_four_bytes_of_sha256_hash,
         constants::WORD_SIZE,
+        to_named,
         types::{
             enum_variants::EnumVariants, errors::Error, param_types::ParamType, StaticStringToken,
             U256,
@@ -494,7 +495,7 @@ mod tests {
         //     x: u32,
         //     y: bool,
         // }
-        let types = vec![ParamType::U32, ParamType::Bool];
+        let types = to_named(&[ParamType::U32, ParamType::Bool]);
         let params = EnumVariants::new(types)?;
 
         // An `EnumSelector` indicating that we've chosen the first Enum variant,
@@ -528,7 +529,7 @@ mod tests {
         // Our enum has two variants: B256, and U64. So the enum will set aside
         // 256b of space or 4 WORDS because that is the space needed to fit the
         // largest variant(B256).
-        let types = vec![ParamType::B256, ParamType::U64];
+        let types = to_named(&[ParamType::B256, ParamType::U64]);
         let enum_variants = EnumVariants::new(types)?;
         let enum_selector = Box::new((1, Token::U64(42), enum_variants));
 
@@ -559,7 +560,7 @@ mod tests {
             v2: str[10]
         }
          */
-        let types = vec![ParamType::Bool, ParamType::StringArray(10)];
+        let types = to_named(&[ParamType::Bool, ParamType::StringArray(10)]);
         let deeper_enum_variants = EnumVariants::new(types)?;
         let deeper_enum_token =
             Token::StringArray(StaticStringToken::new("0123456789".into(), Some(10)));
@@ -571,14 +572,16 @@ mod tests {
         }
          */
 
-        let fields = vec![
+        let fields = to_named(&[
             ParamType::Enum {
-                variants: deeper_enum_variants.clone(),
+                name: "".to_string(),
+                enum_variants: deeper_enum_variants.clone(),
                 generics: vec![],
             },
             ParamType::Bool,
-        ];
+        ]);
         let struct_a_type = ParamType::Struct {
+            name: "".to_string(),
             fields,
             generics: vec![],
         };
@@ -596,7 +599,7 @@ mod tests {
         }
         */
 
-        let types = vec![struct_a_type, ParamType::Bool, ParamType::U64];
+        let types = to_named(&[struct_a_type, ParamType::Bool, ParamType::U64]);
         let top_level_enum_variants = EnumVariants::new(types)?;
         let top_level_enum_token =
             Token::Enum(Box::new((0, struct_a_token, top_level_enum_variants)));
@@ -773,7 +776,7 @@ mod tests {
     fn enums_with_only_unit_variants_are_encoded_in_one_word() -> Result<()> {
         let expected = [0, 0, 0, 0, 0, 0, 0, 1];
 
-        let types = vec![ParamType::Unit, ParamType::Unit];
+        let types = to_named(&[ParamType::Unit, ParamType::Unit]);
         let enum_selector = Box::new((1, Token::Unit, EnumVariants::new(types)?));
 
         let actual = ABIEncoder::default()
@@ -802,7 +805,7 @@ mod tests {
         let padding = vec![0; 32];
         let expected: Vec<u8> = [discriminant, padding].into_iter().flatten().collect();
 
-        let types = vec![ParamType::B256, ParamType::Unit];
+        let types = to_named(&[ParamType::B256, ParamType::Unit]);
         let enum_selector = Box::new((1, Token::Unit, EnumVariants::new(types)?));
 
         let actual = ABIEncoder::default()
@@ -876,7 +879,7 @@ mod tests {
     fn a_vec_in_an_enum() -> Result<()> {
         // arrange
         let offset = 40;
-        let types = vec![ParamType::B256, ParamType::Vector(Box::new(ParamType::U64))];
+        let types = to_named(&[ParamType::B256, ParamType::Vector(Box::new(ParamType::U64))]);
         let variants = EnumVariants::new(types)?;
         let selector = (1, Token::Vector(vec![Token::U64(5)]), variants);
         let token = Token::Enum(Box::new(selector));
@@ -917,7 +920,7 @@ mod tests {
     fn an_enum_in_a_vec() -> Result<()> {
         // arrange
         let offset = 40;
-        let types = vec![ParamType::B256, ParamType::U8];
+        let types = to_named(&[ParamType::B256, ParamType::U8]);
         let variants = EnumVariants::new(types)?;
         let selector = (1, Token::U8(8), variants);
         let enum_token = Token::Enum(Box::new(selector));
@@ -1090,10 +1093,10 @@ mod tests {
         let token = Token::Enum(Box::new((
             1,
             Token::String("".to_string()),
-            EnumVariants::new(vec![
+            EnumVariants::new(to_named(&[
                 ParamType::StringArray(18446742977385549567),
                 ParamType::U8,
-            ])?,
+            ]))?,
         )));
         let capacity_overflow_error = ABIEncoder::default().encode(&[token]).unwrap_err();
 
@@ -1154,7 +1157,7 @@ mod tests {
         let selector = (
             0u64,
             inner_enum,
-            EnumVariants::new(vec![ParamType::U64]).unwrap(),
+            EnumVariants::new(to_named(&[ParamType::U64])).unwrap(),
         );
 
         Token::Enum(Box::new(selector))
