+++ packages/fuels-core/src/types/core/bits.rs
@@ -123,7 +123,8 @@ mod tests {
         assert_eq!(
             EvmAddress::param_type(),
             ParamType::Struct {
-                fields: vec![ParamType::B256],
+                name: "EvmAddress".to_string(),
+                fields: vec![("value".to_string(), ParamType::B256)],
                 generics: vec![]
             }
         );
