+++ packages/fuels-core/src/types/param_types/decodable_debug.rs
@@ -1,0 +1,241 @@
+use std::iter::zip;
+
+use crate::{
+    codec::ABIDecoder,
+    types::{
+        errors::{error, Result},
+        param_types::ParamType,
+        Token,
+    },
+};
+
+impl ParamType {
+    pub fn decode_debug(&self, bytes: &[u8]) -> Result<String> {
+        let token = ABIDecoder::default().decode(self, bytes)?;
+        paramtype_decode_debug(self, &token)
+    }
+}
+
+fn inner_types_debug(tokens: &[Token], inner_type: &ParamType, join_str: &str) -> Result<String> {
+    let inner_types_log = tokens
+        .iter()
+        .map(|token| paramtype_decode_debug(inner_type, token))
+        .collect::<Result<Vec<_>>>()?
+        .join(join_str);
+
+    Ok(inner_types_log)
+}
+
+fn paramtype_decode_debug(param_type: &ParamType, token: &Token) -> Result<String> {
+    let result = match (param_type, token) {
+        (ParamType::Unit, Token::Unit) => "()".to_string(),
+        (ParamType::Bool, Token::Bool(val)) => val.to_string(),
+        (ParamType::U8, Token::U8(val)) => val.to_string(),
+        (ParamType::U16, Token::U16(val)) => val.to_string(),
+        (ParamType::U32, Token::U32(val)) => val.to_string(),
+        (ParamType::U64, Token::U64(val)) => val.to_string(),
+        (ParamType::U128, Token::U128(val)) => val.to_string(),
+        (ParamType::U256, Token::U256(val)) => val.to_string(),
+        (ParamType::B256, Token::B256(val)) => {
+            format!("Bits256({val:?})")
+        }
+        (ParamType::Bytes, Token::Bytes(val)) => {
+            format!("Bytes({val:?})")
+        }
+        (ParamType::String, Token::String(val)) => val.clone(),
+        (ParamType::RawSlice, Token::RawSlice(val)) => {
+            format!("RawSlice({val:?})")
+        }
+        (ParamType::StringArray(..), Token::StringArray(str_token)) => {
+            format!("SizedAsciiString {{ data: \"{}\" }}", str_token.data)
+        }
+        (ParamType::StringSlice, Token::StringSlice(str_token)) => {
+            format!("AsciiString {{ data: \"{}\" }}", str_token.data)
+        }
+        (ParamType::Tuple(types), Token::Tuple(tokens)) => {
+            let elements = zip(types, tokens)
+                .map(|(ptype, token)| paramtype_decode_debug(ptype, token))
+                .collect::<Result<Vec<_>>>()?
+                .join(", ");
+
+            format!("({elements})")
+        }
+        (ParamType::Array(inner_type, _), Token::Array(tokens)) => {
+            let elements = inner_types_debug(tokens, inner_type, ", ")?;
+            format!("[{elements}]")
+        }
+        (ParamType::Vector(inner_type), Token::Vector(tokens)) => {
+            let elements = inner_types_debug(tokens, inner_type, ", ")?;
+            format!("[{elements}]")
+        }
+        (ParamType::Struct { name, fields, .. }, Token::Struct(field_tokens)) => {
+            let fields = zip(fields, field_tokens)
+                .map(|((field_name, param_type), token)| -> Result<_> {
+                    let field_stringified = paramtype_decode_debug(param_type, token)?;
+                    Ok(format!("{field_name}: {field_stringified}"))
+                })
+                .collect::<Result<Vec<_>>>()?
+                .join(", ");
+            format!("{name} {{ {fields} }}")
+        }
+        (ParamType::Enum { .. }, Token::Enum(selector)) => {
+            let (discriminant, token, variants) = selector.as_ref();
+
+            let (variant_name, variant_param_type) = variants.select_variant(*discriminant)?;
+            let variant_str = paramtype_decode_debug(variant_param_type, token)?;
+            let variant_str = if variant_str == "()" {
+                "".into()
+            } else {
+                format!("({variant_str})")
+            };
+
+            format!("{variant_name}{variant_str}")
+        }
+        _ => {
+            return Err(error!(
+                Codec,
+                "could not decode debug from param type: `{param_type:?}` and token: `{token:?}`"
+            ))
+        }
+    };
+    Ok(result)
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        traits::Parameterize,
+        types::{
+            errors::Result, AsciiString, Bits256, Bytes, EvmAddress, RawSlice, SizedAsciiString,
+            U256,
+        },
+    };
+
+    #[test]
+    fn param_type_decode_debug() -> Result<()> {
+        {
+            assert_eq!(
+                format!("{:?}", true),
+                bool::param_type().decode_debug(&[0, 0, 0, 0, 0, 0, 0, 1])?
+            );
+
+            assert_eq!(
+                format!("{:?}", 128u8),
+                u8::param_type().decode_debug(&[0, 0, 0, 0, 0, 0, 0, 128])?
+            );
+
+            assert_eq!(
+                format!("{:?}", 256u16),
+                u16::param_type().decode_debug(&[0, 0, 0, 0, 0, 0, 1, 0])?
+            );
+
+            assert_eq!(
+                format!("{:?}", 512u32),
+                u32::param_type().decode_debug(&[0, 0, 0, 0, 0, 0, 2, 0])?
+            );
+
+            assert_eq!(
+                format!("{:?}", 1024u64),
+                u64::param_type().decode_debug(&[0, 0, 0, 0, 0, 0, 4, 0])?
+            );
+
+            assert_eq!(
+                format!("{:?}", 1024u128),
+                u128::param_type()
+                    .decode_debug(&[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0])?
+            );
+
+            assert_eq!(
+                format!("{:?}", U256::from(2048)),
+                U256::param_type().decode_debug(&[
+                    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                    0, 0, 0, 0, 8, 0
+                ])?
+            );
+        }
+        {
+            let bytes = [
+                239, 134, 175, 169, 105, 108, 240, 220, 99, 133, 226, 196, 7, 166, 225, 89, 161,
+                16, 60, 239, 183, 226, 174, 6, 54, 251, 51, 211, 203, 42, 158, 74,
+            ];
+            let bits256 = Bits256(bytes);
+
+            assert_eq!(
+                format!("{bits256:?}"),
+                Bits256::param_type().decode_debug(&[
+                    239, 134, 175, 169, 105, 108, 240, 220, 99, 133, 226, 196, 7, 166, 225, 89,
+                    161, 16, 60, 239, 183, 226, 174, 6, 54, 251, 51, 211, 203, 42, 158, 74
+                ])?
+            );
+
+            assert_eq!(
+                format!("{:?}", Bytes(bytes.to_vec())),
+                Bytes::param_type().decode_debug(&[
+                    239, 134, 175, 169, 105, 108, 240, 220, 99, 133, 226, 196, 7, 166, 225, 89,
+                    161, 16, 60, 239, 183, 226, 174, 6, 54, 251, 51, 211, 203, 42, 158, 74
+                ])?
+            );
+
+            assert_eq!(
+                format!("{:?}", RawSlice(bytes.to_vec())),
+                RawSlice::param_type().decode_debug(&[
+                    239, 134, 175, 169, 105, 108, 240, 220, 99, 133, 226, 196, 7, 166, 225, 89,
+                    161, 16, 60, 239, 183, 226, 174, 6, 54, 251, 51, 211, 203, 42, 158, 74
+                ])?
+            );
+
+            assert_eq!(
+                format!("{:?}", EvmAddress::from(bits256)),
+                EvmAddress::param_type().decode_debug(&[
+                    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 166, 225, 89, 161, 16, 60, 239, 183,
+                    226, 174, 6, 54, 251, 51, 211, 203, 42, 158, 74
+                ])?
+            );
+        }
+        {
+            assert_eq!(
+                format!("{:?}", AsciiString::new("Fuel".to_string())?),
+                AsciiString::param_type().decode_debug(&[70, 117, 101, 108])?
+            );
+
+            assert_eq!(
+                format!("{:?}", SizedAsciiString::<4>::new("Fuel".to_string())?),
+                SizedAsciiString::<4>::param_type()
+                    .decode_debug(&[70, 117, 101, 108, 0, 0, 0, 0])?
+            );
+
+            assert_eq!(
+                format!("{}", "Fuel"),
+                String::param_type().decode_debug(&[70, 117, 101, 108])?
+            );
+        }
+        {
+            assert_eq!(
+                format!("{:?}", (1, 2)),
+                <(u8, u8)>::param_type()
+                    .decode_debug(&[1, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0])?
+            );
+
+            assert_eq!(
+                format!("{:?}", [3, 4]),
+                <[u64; 2]>::param_type()
+                    .decode_debug(&[0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0, 4])?
+            );
+        }
+        {
+            assert_eq!(
+                format!("{:?}", Some(42)),
+                <Option<u64>>::param_type()
+                    .decode_debug(&[0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 42])?
+            );
+
+            assert_eq!(
+                format!("{:?}", Err::<u64, u64>(42u64)),
+                <std::result::Result<u64, u64>>::param_type()
+                    .decode_debug(&[0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 42])?
+            );
+        }
+
+        Ok(())
+    }
+}
