+++ packages/fuels-macros/src/derive/parameterize.rs
@@ -44,7 +44,9 @@ fn parameterize_for_struct(
     no_std: bool,
 ) -> Result<TokenStream> {
     let (impl_gen, type_gen, where_clause) = generics.split_for_impl();
+    let name_stringified = name.to_string();
     let members = Members::from_struct(contents, fuels_core_path.clone())?;
+    let field_names = members.names_as_strings();
     let param_type_calls = members.param_type_calls();
     let generic_param_types = parameterize_generic_params(&generics, &fuels_core_path)?;
 
@@ -54,7 +56,8 @@ fn parameterize_for_struct(
         impl #impl_gen #fuels_core_path::traits::Parameterize for #name #type_gen #where_clause {
             fn param_type() -> #fuels_types_path::param_types::ParamType {
                 #fuels_types_path::param_types::ParamType::Struct{
-                    fields: #std_lib::vec![#(#param_type_calls),*],
+                    name: #std_lib::string::String::from(#name_stringified),
+                    fields: #std_lib::vec![#((#field_names, #param_type_calls)),*],
                     generics: #std_lib::vec![#(#generic_param_types),*],
                 }
             }
@@ -89,6 +92,7 @@ fn parameterize_for_enum(
     let enum_name_str = name.to_string();
     let members = Members::from_enum(contents, fuels_core_path.clone())?;
 
+    let variant_names = members.names_as_strings();
     let variant_param_types = members.param_type_calls();
     let generic_param_types = parameterize_generic_params(&generics, &fuels_core_path)?;
 
@@ -97,17 +101,17 @@ fn parameterize_for_enum(
     Ok(quote! {
         impl #impl_gen #fuels_core_path::traits::Parameterize for #name #type_gen #where_clause {
             fn param_type() -> #fuels_types_path::param_types::ParamType {
-                let variants = #std_lib::vec![#(#variant_param_types),*];
-
-                let variants = #fuels_types_path::enum_variants::EnumVariants::new(variants)
+                let variants = #std_lib::vec![#((#variant_names, #variant_param_types)),*];
+                let enum_variants = #fuels_types_path::enum_variants::EnumVariants::new(variants)
                     .unwrap_or_else(|_| ::std::panic!(
-                            "{} has no variants which isn't allowed.",
+                            "{} has no variants which isn't allowed",
                             #enum_name_str
                         )
                     );
 
                 #fuels_types_path::param_types::ParamType::Enum {
-                    variants,
+                    name: #std_lib::string::String::from(#enum_name_str),
+                    enum_variants,
                     generics: #std_lib::vec![#(#generic_param_types),*]
                 }
             }
